# openapi.yaml
openapi: 3.0.3
info:
  title: DOGFOOD API
  version: 0.1.0
  description: 'Hackathon judging, voting, and management API. Note that admin/service routes typically wrap responses in { data: ... } while some public routes return unwrapped JSON. Error formats also vary between wrapped structured objects and raw string messages.'
servers:
  - url: 'http://localhost:3000'
    description: Local deployment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  '/api/hackathons/{id}/webhooks':
    get:
      summary: List Webhooks
      description: Returns a list of webhooks for the hackathon. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
    post:
      summary: Create Webhook
      description: Creates a new webhook endpoint. Returns the secret once. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
              required:
                - url
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      secret:
                        type: string
  '/api/hackathons/{id}/webhooks/{webhookId}':
    get:
      summary: Get Webhook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: webhookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
    put:
      summary: Update Webhook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: webhookId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                is_active:
                  type: boolean
              required:
                - url
                - is_active
      responses:
        '200':
          description: OK
    delete:
      summary: Delete Webhook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: webhookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
  '/api/hackathons/{id}/webhooks/{webhookId}/test':
    post:
      summary: Send Test Webhook
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: webhookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...